Simple overdraw test app
This app draws successive nearly-full-screen rectangles
to see the impact of additional overdraw on performance.
Performance is shown as a moving average of the previous
ten frames.

The rectangles are drawn with 50% alpha in an effort to
ensure that none of the writes are optimized out, and the
total number draw calls (200, at present) is kept
constant to eliminate draw call overhead from the
measurement.